--- src/helpers/convertGlobPaths.js
@@ -1,14 +1,0 @@
-const glob = require('glob');
-
-/**
- * Converts an array of globs to full paths
- * @param {array} globs - Array of globs and/or normal paths
- * @return {array} Array of fully-qualified paths
- */
-function convertGlobPaths(globs) {
-  return globs
-    .map((globString) => glob.sync(globString))
-    .reduce((previous, current) => previous.concat(current), []);
-}
-
-module.exports = convertGlobPaths;
