+++ test/utils.spec.js
@@ -1,97 +1,8 @@
 /* eslint no-unused-expressions: 0 */
-const specHelper = require('../src/helpers/specification');
-const hasEmptyProperty = require('../src/helpers/hasEmptyProperty');
-const parseApiFileContent = require('../src/helpers/parseApiFileContent');
 
-const swaggerObject = require('./files/v2/swaggerObject.json');
-const testData = require('./files/v2/testData');
+const utils = require('../src/utils');
 
-describe('Helpers', () => {
-  describe('addDataToSwaggerObject', () => {
-    it('should be a function', () => {
-      expect(typeof specHelper.addDataToSwaggerObject).toBe('function');
-    });
-
-    it('should validate input', () => {
-      expect(() => {
-        specHelper.addDataToSwaggerObject();
-      }).toThrow('swaggerObject and data are required!');
-    });
-
-    it('should handle "definition" and "definitions"', () => {
-      specHelper.addDataToSwaggerObject(swaggerObject, testData.definitions);
-      expect(swaggerObject.definitions).toEqual({
-        DefinitionSingular: {
-          required: ['username', 'password'],
-          properties: {
-            username: { type: 'string' },
-            password: { type: 'string' },
-          },
-        },
-        DefinitionPlural: {
-          required: ['username', 'password'],
-          properties: {
-            username: { type: 'string' },
-            password: { type: 'string' },
-          },
-        },
-      });
-    });
-
-    it('should handle "parameter" and "parameters"', () => {
-      specHelper.addDataToSwaggerObject(swaggerObject, testData.parameters);
-      expect(swaggerObject.parameters).toEqual({
-        ParameterSingular: {
-          name: 'username',
-          description: 'Username to use for login.',
-          in: 'formData',
-          required: true,
-          type: 'string',
-        },
-        ParameterPlural: {
-          name: 'limit',
-          in: 'query',
-          description: 'max records to return',
-          required: true,
-          type: 'integer',
-          format: 'int32',
-        },
-      });
-    });
-
-    it('should handle "securityDefinition" and "securityDefinitions"', () => {
-      specHelper.addDataToSwaggerObject(
-        swaggerObject,
-        testData.securityDefinitions
-      );
-      expect(swaggerObject.securityDefinitions).toEqual({
-        basicAuth: {
-          type: 'basic',
-          description:
-            'HTTP Basic Authentication. Works over `HTTP` and `HTTPS`',
-        },
-        api_key: { type: 'apiKey', name: 'api_key', in: 'header' },
-        petstore_auth: {
-          type: 'oauth2',
-          authorizationUrl: 'http://swagger.io/api/oauth/dialog',
-          flow: 'implicit',
-          scopes: {
-            'write:pets': 'modify pets in your account',
-            'read:pets': 'read your pets',
-          },
-        },
-      });
-    });
-
-    it('should handle "response" and "responses"', () => {
-      specHelper.addDataToSwaggerObject(swaggerObject, testData.responses);
-      expect(swaggerObject.responses).toEqual({
-        NotFound: { description: 'Entity not found.' },
-        IllegalInput: { description: 'Illegal input for operation.' },
-      });
-    });
-  });
-
+describe('Utilities module', () => {
   describe('hasEmptyProperty', () => {
     it('identifies object with an empty object or array as property', () => {
       const invalidA = { foo: {} };
@@ -100,11 +11,11 @@ describe('Helpers', () => {
       const validB = { foo: ['¯_(ツ)_/¯'] };
       const validC = { foo: '¯_(ツ)_/¯' };
 
-      expect(hasEmptyProperty(invalidA)).toBe(true);
-      expect(hasEmptyProperty(invalidB)).toBe(true);
-      expect(hasEmptyProperty(validA)).toBe(false);
-      expect(hasEmptyProperty(validB)).toBe(false);
-      expect(hasEmptyProperty(validC)).toBe(false);
+      expect(utils.hasEmptyProperty(invalidA)).toBe(true);
+      expect(utils.hasEmptyProperty(invalidB)).toBe(true);
+      expect(utils.hasEmptyProperty(validA)).toBe(false);
+      expect(utils.hasEmptyProperty(validB)).toBe(false);
+      expect(utils.hasEmptyProperty(validC)).toBe(false);
     });
   });
 
@@ -140,7 +51,7 @@ describe('Helpers', () => {
         };
       `;
 
-      expect(parseApiFileContent(fileContent, '.js')).toEqual({
+      expect(utils.parseApiFileContent(fileContent, '.js')).toEqual({
         yaml: [],
         jsdoc: [
           {
@@ -197,7 +108,7 @@ describe('Helpers', () => {
         res.json req.body
     `;
 
-      expect(parseApiFileContent(fileContent, '.coffee')).toEqual({
+      expect(utils.parseApiFileContent(fileContent, '.coffee')).toEqual({
         yaml: [],
         jsdoc: [
           {
